refactor(rust): expose the Cli struct from agama-cli
